requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary    # PostgreSQL
pillow             # images
weasyprint         # invoice PDF
requests           # PhonePe API + WhatsApp API
